Fix GDA settings
