+++ auto/src/main/res/xml/pref_general.xml
@@ -8,16 +8,32 @@
             android:title="@string/pref_switch_use_mmol"
             app:iconSpaceReserved="false" />
         <androidx.preference.SwitchPreferenceCompat
-            android:defaultValue="true"
-            android:key="relative_time"
-            android:summary="@string/pref_relative_time_summary"
-            android:title="@string/pref_relative_time"
-            app:iconSpaceReserved="false" />
-        <androidx.preference.SwitchPreferenceCompat
             android:defaultValue="false"
             android:key="five_minute_delta"
             android:summary="@string/pref_switch_use_5_min_delta_summary"
             android:title="@string/pref_switch_use_5_min_delta"
+            app:iconSpaceReserved="false" />
+        <EditTextPreference
+            android:defaultValue=""
+            android:key="patient_name"
+            android:title="@string/patient_name"
+            android:summary="@string/patient_name_gda_summary"
+            app:iconSpaceReserved="false"/>
+        <PreferenceCategory
+            android:title="@string/advanced_settings"
+            app:iconSpaceReserved="false">
+        <androidx.preference.SwitchPreferenceCompat
+            android:defaultValue="false"
+            android:key="rate_calculation"
+            android:summaryOn="@string/pref_switch_use_rate_calculation_summary_on"
+            android:summaryOff="@string/pref_switch_use_rate_calculation_summary_off"
+            android:title="@string/pref_switch_use_rate_calculation"
+            app:iconSpaceReserved="false" />
+        <androidx.preference.SwitchPreferenceCompat
+            android:defaultValue="true"
+            android:key="relative_time"
+            android:summary="@string/pref_relative_time_summary"
+            android:title="@string/pref_relative_time"
             app:iconSpaceReserved="false" />
         <androidx.preference.SeekBarPreference
             android:defaultValue="6"
@@ -29,14 +45,5 @@
             android:title="@string/obsolete_time"
             android:summary="@string/obsolete_time_summary"
             app:iconSpaceReserved="false" />
-        <PreferenceCategory
-            android:title="@string/pref_cat_android_auto"
-            app:iconSpaceReserved="false">
-                <EditTextPreference
-                    android:defaultValue=""
-                    android:key="patient_name"
-                    android:title="@string/patient_name"
-                    android:summary="@string/patient_name_gda_summary"
-                    app:iconSpaceReserved="false"/>
         </PreferenceCategory>
 </PreferenceScreen>
